requirements.txt
# openai
# langchain
# faiss-cpu
# pymupdf
# tiktoken
# streamlit
# python-dotenv


# Backend
fastapi
uvicorn
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose
openai
langchain
chromadb

# Frontend
streamlit
requests

faiss-cpu
tiktoken
